third_party/blink/web_tests/virtual/threaded-prefer-compositing/fast/scroll-behavior/smooth-scroll/mousewheel-scroll-interrupted-expected.txt

This test ensures that input driven user scrolls get cancelled by non-input driven scrolls.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.getScrollAnimationState(document) is RUNNING_ON_COMPOSITOR
PASS isOrWillBeCancelled(document) is true
PASS document.scrollingElement.scrollTop < 40 && document.scrollingElement.scrollLeft == 0 became true
PASS successfullyParsed is true

TEST COMPLETE
